src/event-sourcer-proto/requirements.txt
protobuf==3.6.0
google-cloud==0.33.1